Include segment_images.py in the cuisto package * seg module is renamed to segmentation * functions from the segment_images.py script moved to the segmentation module * the script moved to the examples * bumped version (will be released to PyPI later on)
