+++ cuisto/segmentation.py
@@ -1,19 +1,24 @@
-"""seg module, part of cuisto.
+"""segmentation module, part of cuisto.
 
 Functions for segmentating probability map stored as an image.
 
 """
 
+import datetime
+import os
 import uuid
+from pathlib import Path
 
 import geojson
 import numpy as np
 import pandas as pd
 import shapely
 import tifffile
+import tqdm
 from skan import Skeleton, summarize
 from skimage import measure, morphology
 
+pd.options.mode.copy_on_write = True  # prepare for pandas 3
 
 def get_pixelsize(image_name: str) -> float:
     """
@@ -444,3 +449,341 @@ def segment_points(
     return get_collection_from_points(
         coords, geojson_props, rescale_factor=rescale_factor
     )
+
+
+def get_seg_method(segtype: str):
+    """
+    Determine what kind of segmentation is performed.
+
+    Segmentation kind are, for now, lines, polygons or points. We detect that based on
+    hardcoded keywords.
+
+    Parameters
+    ----------
+    segtype : str
+
+    Returns
+    -------
+    seg_method : str
+
+    """
+
+    line_list = ["fibers", "axons", "fiber", "axon"]
+    point_list = ["synapto", "synaptophysin", "syngfp", "boutons", "points"]
+    polygon_list = ["cells", "polygon", "polygons", "polygon", "cell"]
+
+    if segtype in line_list:
+        seg_method = "lines"
+    elif segtype in polygon_list:
+        seg_method = "polygons"
+    elif segtype in point_list:
+        seg_method = "points"
+    else:
+        raise ValueError(
+            f"Could not determine method to use based on segtype : {segtype}."
+        )
+
+    return seg_method
+
+
+def get_geojson_dir(images_dir: str):
+    """
+    Get the directory of geojson files, which will be in the parent directory
+    of `images_dir`.
+
+    If the directory does not exist, create it.
+
+    Parameters
+    ----------
+    images_dir : str
+
+    Returns
+    -------
+    geojson_dir : str
+
+    """
+
+    geojson_dir = os.path.join(Path(images_dir).parent, "geojson")
+
+    if not os.path.isdir(geojson_dir):
+        os.mkdir(geojson_dir)
+
+    return geojson_dir
+
+
+def get_geojson_properties(name: str, color: tuple | list, objtype: str = "detection"):
+    """
+    Return geojson objects properties as a dictionnary, ready to be used in geojson.Feature.
+
+    Parameters
+    ----------
+    name : str
+        Classification name.
+    color : tuple or list
+        Classification color in RGB (3-elements vector).
+    objtype : str, optional
+        Object type ("detection" or "annotation"). Default is "detection".
+
+    Returns
+    -------
+    props : dict
+
+    """
+
+    return {
+        "objectType": objtype,
+        "classification": {"name": name, "color": color},
+        "isLocked": "true",
+    }
+
+
+def parameters_as_dict(
+    images_dir: str,
+    masks_dir: str,
+    segtype: str,
+    name: str,
+    proba_threshold: float,
+    edge_dist: float,
+):
+    """
+    Get information as a dictionnary.
+
+    Parameters
+    ----------
+    images_dir : str
+        Path to images to be segmented.
+    masks_dir : str
+        Path to images masks.
+    segtype : str
+        Segmentation type (eg. "fibers").
+    name : str
+        Name of the segmentation (eg. "green").
+    proba_threshold : float < 1
+        Probability threshold.
+    edge_dist : float
+        Distance in µm to the brain edge that is ignored.
+
+    Returns
+    -------
+    params : dict
+
+    """
+
+    return {
+        "images_location": images_dir,
+        "masks_location": masks_dir,
+        "type": segtype,
+        "probability threshold": proba_threshold,
+        "name": name,
+        "edge distance": edge_dist,
+    }
+
+
+def write_parameters(
+    outfile: str, parameters: dict, filters: dict, original_pixelsize: float
+):
+    """
+    Write parameters to `outfile`.
+
+    A timestamp will be added. Parameters are written as key = value,
+    and a [filters] is added before filters parameters.
+
+    Parameters
+    ----------
+    outfile : str
+        Full path to the output file.
+    parameters : dict
+        General parameters.
+    filters : dict
+        Filters parameters.
+    original_pixelsize : float
+        Size of pixels in original image.
+
+    """
+
+    with open(outfile, "w") as fid:
+        fid.writelines(f"date = {datetime.now().strftime('%d-%B-%Y %H:%M:%S')}\n")
+
+        fid.writelines(f"original_pixelsize = {original_pixelsize}\n")
+
+        for key, value in parameters.items():
+            fid.writelines(f"{key} = {value}\n")
+
+        fid.writelines("[filters]\n")
+
+        for key, value in filters.items():
+            fid.writelines(f"{key} = {value}\n")
+
+
+def process_directory(
+    images_dir: str,
+    img_suffix: str = "",
+    segtype: str = "",
+    original_pixelsize: float = 1.0,
+    target_channel: int = 0,
+    proba_threshold: float = 0.0,
+    max_pix_value: float = 255,
+    qupath_class: str = "Object",
+    qupath_color: list = [0, 0, 0],
+    qupath_type: str = "Detection",
+    channel_suffix: str = "",
+    edge_dist: float = 0.0,
+    filters: dict = {},
+    masks_dir: str = "",
+    masks_ext: str = "",
+):
+    """
+    Segment the .ome.tiff files in the input directory.
+
+    Parameters
+    ----------
+    images_dir : str
+        Animal ID to process.
+    img_suffix : str
+        Images suffix, including extension.
+    segtype : str
+        Segmentation type.
+    original_pixelsize : float
+        Original images pixel size in microns.
+    target_channel : int
+        Index of the channel containning the objects of interest (eg. not the
+        background), in the probability map (*not* the original images channels).
+    proba_threshold : float < 1
+        Probability below this value will be discarded (multiplied by `max_pixel_value`)
+    max_pixel_value : float
+        Maximum pixel value in the segmented image, to rescale them to a probability
+        between 0 and 1.
+    qupath_class : str
+        Name of the QuPath classification.
+    qupath_color : list of three elements
+        Color associated to that classification in RGB.
+    qupath_type : {"Detection", "Annotation"}
+        QuPath type of object.
+    channel_suffix : str
+        Channel name, will be used as a suffix in output geojson files.
+    edge_dist : float
+        Distance to the edge of the brain masks that will be ignored, in microns. Set to
+        0 to disable this feature.
+    filters : dict
+        Filters values to include or excludes objects. See the top of the script.
+    masks_dir : str, optional
+        Path to images masks, to exclude objects found near the edges. The masks must be
+        with the same name as the corresponding image to be segmented, without its
+        suffix. Default is "", which disables this feature.
+    masks_ext : str, optional
+        Masks files extension, without leading ".". Default is ""
+
+    """
+
+    # -- Preparation
+    # get segmentation type
+    seg_method = get_seg_method(segtype)
+
+    # get output directory path
+    geojson_dir = get_geojson_dir(images_dir)
+
+    # get images list
+    images_list = [
+        os.path.join(images_dir, filename)
+        for filename in os.listdir(images_dir)
+        if filename.endswith(img_suffix)
+    ]
+
+    if len(images_list) == 0:
+        raise FileNotFoundError(
+            f"No file found in {images_dir}. Check 'IMAGES_DIR' and 'IMG_SUFFIX'."
+        )
+
+    # write parameters
+    parameters = parameters_as_dict(
+        images_dir, masks_dir, segtype, channel_suffix, proba_threshold, edge_dist
+    )
+    param_file = os.path.join(geojson_dir, "parameters" + channel_suffix + ".txt")
+    if os.path.isfile(param_file):
+        raise FileExistsError("Parameters file already exists.")
+    else:
+        write_parameters(param_file, parameters, filters, original_pixelsize)
+
+    # convert parameters to pixels in probability map
+    pixelsize = get_pixelsize(images_list[0])  # get pixel size
+    edge_dist = int(edge_dist / pixelsize)
+    filters = convert_to_pixels(filters, pixelsize)
+
+    # get rescaling factor
+    rescale_factor = pixelsize / original_pixelsize
+
+    # get GeoJSON properties
+    geojson_props = get_geojson_properties(
+        qupath_class, qupath_color, objtype=qupath_type
+    )
+
+    # -- Processing
+    pbar = tqdm(images_list)
+    for imgpath in pbar:
+        # build file names
+        imgname = os.path.basename(imgpath)
+        geoname = imgname.replace(img_suffix, "")
+        geojson_file = os.path.join(
+            geojson_dir, geoname + "_segmentation" + channel_suffix + ".geojson"
+        )
+
+        # checks if output file already exists
+        if os.path.isfile(geojson_file):
+            continue
+
+        # read images
+        pbar.set_description(f"{geoname}: Loading...")
+        img = tifffile.imread(imgpath, key=target_channel)
+        if (edge_dist > 0) & (len(masks_dir) != 0):
+            mask = tifffile.imread(os.path.join(masks_dir, geoname + "." + masks_ext))
+            mask = pad_image(mask, img.shape)  # resize mask
+            # apply mask, eroding from the edges
+            img = img * erode_mask(mask, edge_dist)
+
+        # image processing
+        pbar.set_description(f"{geoname}: IP...")
+
+        # threshold probability and binarization
+        img = img >= proba_threshold * max_pix_value
+
+        # segmentation
+        pbar.set_description(f"{geoname}: Segmenting...")
+
+        if seg_method == "lines":
+            collection = segment_lines(
+                img,
+                geojson_props,
+                minsize=filters["length_low"],
+                rescale_factor=rescale_factor,
+            )
+
+        elif seg_method == "polygons":
+            collection = segment_polygons(
+                img,
+                geojson_props,
+                area_min=filters["area_low"],
+                area_max=filters["area_high"],
+                ecc_min=filters["ecc_low"],
+                ecc_max=filters["ecc_high"],
+                rescale_factor=rescale_factor,
+            )
+
+        elif seg_method == "points":
+            collection = segment_points(
+                img,
+                geojson_props,
+                area_min=filters["area_low"],
+                area_max=filters["area_high"],
+                ecc_min=filters["ecc_low"],
+                ecc_max=filters["ecc_high"],
+                dist_thresh=filters["dist_thresh"],
+                rescale_factor=rescale_factor,
+            )
+        else:
+            # we already printed an error message
+            return
+
+        # save geojson
+        pbar.set_description(f"{geoname}: Saving...")
+        with open(geojson_file, "w") as fid:
+            fid.write(geojson.dumps(collection))
